requirements.txt
crewai==1.7.1
python-dotenv
pydantic
